Added code to ArquillianSuiteExtension to determine if being run in a Suite
We wanted to be able to run single test with a smaller deployment
defined only for that test.  This change then will use @Deploymnet
defined in the test when not part of suite, and if being run as part
of a suite it will then look for @ArquillianSuiteDeployment for
deployments.

Any test that provides it's own @Deployment should not extend
@ArquillianSuiteDeployment since that will cause an error when
run as part of a suite.