Bug? Duplicated class annotated with ArquillianS..
Is it a bug? @ArquillianSuiteDeployment is duplicated?
Fixing a bug when inside eclipse you open two or more projects that are dependency with other and it shows that the @ArquillianSuiteDeployment is duplicated. Maybe it happens on maven test also.
Example, can be project or module:
Project A
Project B
Project A - Has a Test Class with @ArquillianSuiteDeployment and @Deployment
Project B - Depends on Project A
Project B - Has your own Test Classes with @ArquillianSuiteDeployment and @Deployment 
Project B - mvn test or run test inside eclipse.
Project B - log.log(Level.SEVERE, "arquillian-suite-deployment: Duplicated class annotated with @ArquillianSuiteDeployment: {0}", type.getName());

Is it normal? Because using that code that I'm pushing solved my problem and my projects runs smoothly and at light speed.
I found another problem that when the application is undeploying, it doesn't undeploy the datasource, when the second project will run, it says that the datasource already exists. But it can be another change, I'll search something.
If it's not the right place to commit, sorry, just say to me where I should commit.

Thank you!